AdminWeb SignUp commit 1.1

+++ DYCampmate/src/main/java/com/Campmate/DYCampmate/repository/AdminRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AdminRepo extends JpaRepository<AdminEntity, Integer> {
+public interface AdminRepo extends JpaRepository<AdminEntity, Long> {
 
     Optional<AdminEntity> findByEmail(String email);
     boolean existsByEmail(String email);
